improvements to operator
check hilbert indexable in LocalOperator
make dtype abstract property of all operators
make get_conn a default implementation in abstractoperator
use dtype in special hamil
