+++ netket/operator/__init__.py
@@ -1,31 +1,25 @@
-from ._local_values import local_values
-
-from ._der_local_values import der_local_values
-
+from ._abstract_operator import AbstractOperator
 
 from ._local_operator import LocalOperator
-from ._local_liouvillian import AbstractSuperOperator, LocalLiouvillian
 from ._graph_operator import GraphOperator
+from ._pauli_strings import PauliStrings
+from ._lazy import Adjoint, Transpose, Squared
+from ._hamiltonian import Ising, Heisenberg, BoseHubbard
+
+from ._local_liouvillian import AbstractSuperOperator, LocalLiouvillian
 
 from . import spin, boson
 
-from ._hamiltonian import Ising, Heisenberg
+# TODO: Deprecated. Remove in v3.1
+from ._local_values import local_values
+from ._der_local_values import der_local_values
+from ._der_local_values_jax import local_energy_kernel
 
-from ._abstract_operator import AbstractOperator
-from ._bose_hubbard import BoseHubbard
-from ._pauli_strings import PauliStrings
-
-from ._lazy import Adjoint, Transpose, Squared
-
-from netket.utils import jax_available as _jax_available
-
-if _jax_available:
-    from ._local_cost_functions import (
-        define_local_cost_function,
-        local_cost_function,
-        local_cost_and_grad_function,
-        local_costs_and_grads_function,
-        local_value_cost,
-        local_value_op_op_cost,
-    )
-    from ._der_local_values_jax import local_energy_kernel
+from ._local_cost_functions import (
+    define_local_cost_function,
+    local_cost_function,
+    local_cost_and_grad_function,
+    local_costs_and_grads_function,
+    local_value_cost,
+    local_value_op_op_cost,
+)
