+++ netket/operator/_pauli_strings.py
@@ -11,7 +11,7 @@ from ._abstract_operator import AbstractOperator
 class PauliStrings(AbstractOperator):
     """A Hamiltonian consisiting of the sum of products of Pauli operators."""
 
-    def __init__(self, operators, weights, cutoff=1.0e-10):
+    def __init__(self, operators, weights, cutoff=1.0e-10, dtype=complex):
         """
         Constructs a new ``PauliStrings`` operator given a set of Pauli operators.
 
@@ -134,40 +134,13 @@ class PauliStrings(AbstractOperator):
         self._z_check = _z_check
 
         self._x_prime_max = np.empty((n_operators, _n_qubits))
-        self._mels_max = np.empty((n_operators), dtype=np.complex128)
+        self._mels_max = np.empty((n_operators), dtype=dtype)
         self._n_operators = n_operators
+        self._dtype = dtype
 
-    def get_conn(self, x):
-        r"""Finds the connected elements of the Operator. Starting
-        from a given quantum number x, it finds all other quantum numbers x' such
-        that the matrix element :math:`O(x,x')` is different from zero. In general there
-        will be several different connected states x' satisfying this
-        condition, and they are denoted here :math:`x'(k)`, for :math:`k=0,1...N_{\mathrm{connected}}`.
-
-        This is a batched version, where x is a matrix of shape (batch_size,hilbert.size).
-
-        Args:
-            x (array): An array of shape (hilbert.size) containing the quantum numbers x.
-
-        Returns:
-            matrix: The connected states x' of shape (N_connected,hilbert.size)
-            array: An array containing the matrix elements :math:`O(x,x')` associated to each x'.
-
-        """
-        return self._flattened_kernel(
-            np.atleast_2d(x),
-            self._x_prime_max,
-            np.ones(1),
-            self._mels_max,
-            self._sites,
-            self._ns,
-            self._n_op,
-            self._weights,
-            self._nz_check,
-            self._z_check,
-            self._cutoff,
-            self._n_operators,
-        )
+    @property
+    def dtype(self):
+        return self._dtype
 
     @staticmethod
     @jit(nopython=True)
@@ -187,7 +160,7 @@ class PauliStrings(AbstractOperator):
     ):
 
         x_prime = np.empty((x.shape[0] * max_conn, x_prime.shape[1]))
-        mels = np.empty((x.shape[0] * max_conn), dtype=np.complex128)
+        mels = np.empty((x.shape[0] * max_conn), dtype=mels.dtype)
 
         n_c = 0
         for b in range(x.shape[0]):
